+++ src/test/psdReader.spec.ts
@@ -69,7 +69,7 @@ describe('PsdReader', () => {
 
 	// skipping "pattern" test because it requires zip cimpression of patterns
 	fs.readdirSync(readFilesPath).filter(f => !/pattern/.test(f)).forEach(f => {
-		// fs.readdirSync(readFilesPath).filter(f => /psb-x/.test(f)).forEach(f => {
+		// fs.readdirSync(readFilesPath).filter(f => /text-bug/.test(f)).forEach(f => {
 		it(`reads PSD file (${f})`, () => {
 			const basePath = path.join(readFilesPath, f);
 			const fileName = fs.existsSync(path.join(basePath, 'src.psb')) ? 'src.psb' : 'src.psd';
@@ -139,7 +139,7 @@ describe('PsdReader', () => {
 	});
 
 	fs.readdirSync(readWriteFilesPath).forEach(f => {
-		// fs.readdirSync(readWriteFilesPath).filter(f => /werps/.test(f)).forEach(f => {
+		// fs.readdirSync(readWriteFilesPath).filter(f => /annot/.test(f)).forEach(f => {
 		it(`reads-writes PSD file (${f})`, () => {
 			const ext = fs.existsSync(path.join(readWriteFilesPath, f, 'src.psb')) ? 'psb' : 'psd';
 			const psd = readPsdFromFile(path.join(readWriteFilesPath, f, `src.${ext}`), {
@@ -155,7 +155,7 @@ describe('PsdReader', () => {
 			// fs.writeFileSync('temp.txt', require('util').inspect(psd, false, 99, false), 'utf8');
 			// fs.writeFileSync('temp2.txt', require('util').inspect(psd2, false, 99, false), 'utf8');
 
-			compareBuffers(actual, expected, `read-write-${f}`, 0x81e10);
+			compareBuffers(actual, expected, `read-write-${f}`, 0);
 		});
 	});